ERA5 reader stuff

+++ dnora/bnd/read_ec.py
@@ -34,18 +34,18 @@ def download_era5_from_cds(start_time, end_time, lon, lat, dlon, dlat, folder='d
 
     filename = f'{folder}/EC_ERA5.nc'
 
-    years = [f'{y:4.0f}' for y in int_list_of_years(start_time, end_time)]
-    months = [f'{m:02.0f}' for m in int_list_of_months(start_time, end_time)]
-    days = [f'{d:02.0f}' for d in int_list_of_days(start_time, end_time)]
+    # years = [f'{y:4.0f}' for y in int_list_of_years(start_time, end_time)]
+    # months = [f'{m:02.0f}' for m in int_list_of_months(start_time, end_time)]
+    # days = [f'{d:02.0f}' for d in int_list_of_days(start_time, end_time)]
 
     # Create string for dates
-    dates = []
-    for y in years:
-        for m in months:
-            for d in days:
-                dates.append(f'{y}-{m}-{d}')
-    dates = '/'.join(dates)
-
+    #dates = []
+    #for y in years:
+    #    for m in months:
+    #        for d in days:
+    #            dates.append(f'{y}-{m}-{d}')
+    #dates = '/'.join(dates)
+    dates = f'{str(start_time)[0:10]}/to/{str(end_time)[0:10]}'
     # cds_command = {
     # 'class': 'ea',
     # 'date': '2018-01-01/to/2018-01-01',
@@ -62,31 +62,72 @@ def download_era5_from_cds(start_time, end_time, lon, lat, dlon, dlat, folder='d
     # 'format': 'netcdf',
     # }
 
-    cds_command ={
+# This worked 9.11.2022
+# cds_command = {
+#     'class': 'ea',
+#     'date': '2018-01-01/to/2018-01-02',
+#     'direction': '1/2/3/4/5/6/7/8/9/10/11/12/13/14/15/16/17/18/19/20/21/22/23/24',
+#     'domain': 'g',
+#     'area': '65/-10/50/5',
+#     'grid': '5/5',
+#     'expver': '1',
+#     'frequency': '1/2/3/4/5/6/7/8/9/10/11/12/13/14/15/16/17/18/19/20/21/22/23/24/25/26/27/28/29/30',
+#     'param': '251.140',
+#     'stream': 'wave',
+#     'time': '00:00:00/01:00:00/02:00:00/03:00:00/04:00:00/05:00:00/06:00:00/07:00:00/08:00:00/09:00:00/10:00:00/11:00:00/12:00:00/13:00:00/14:00:00/15:00:00/16:00:00/17:00:00/18:00:00/19:00:00/20:00:00/21:00:00/22:00:00/23:00:00',
+#     'type': 'an',
+#     'format': 'netcdf',
+# }
+
+    cds_command = {
         'class': 'ea',
         'date': dates,
-        'direction': '/'.join([f'{n+1:01.0f}' for n in range(24)]),
+        'direction': '1/2/3/4/5/6/7/8/9/10/11/12/13/14/15/16/17/18/19/20/21/22/23/24',
         'domain': 'g',
+        'area': f'{lat[1]}/{lon[0]}/{lat[0]}/{lon[1]}', # north, west, south, east
+        'grid': f'{dlat}/{dlon}',
         'expver': '1',
-        'frequency': '/'.join([f'{n+1:01.0f}' for n in range(30)]),
+        'frequency': '1/2/3/4/5/6/7/8/9/10/11/12/13/14/15/16/17/18/19/20/21/22/23/24/25/26/27/28/29/30',
         'param': '251.140',
         'stream': 'wave',
-        'time': '/'.join([f'{n:02.0f}:00:00' for n in range(24)]),
-        'area': f'{lat[1]+0.0001}/{lon[0]}/{lat[0]}/{lon[1]+0.0001}', # north, west, south, east
-        'grid': f'{dlon}/{dlat}',
+        'time': '00:00:00/01:00:00/02:00:00/03:00:00/04:00:00/05:00:00/06:00:00/07:00:00/08:00:00/09:00:00/10:00:00/11:00:00/12:00:00/13:00:00/14:00:00/15:00:00/16:00:00/17:00:00/18:00:00/19:00:00/20:00:00/21:00:00/22:00:00/23:00:00',
         'type': 'an',
         'format': 'netcdf',
-        }
+    }
+    print(cds_command)
+
+    # cds_command ={
+    #     'class': 'ea',
+    #     'date': dates,
+    #     'direction': '/'.join([f'{n+1:01.0f}' for n in range(24)]),
+    #     'domain': 'g',
+    #     'expver': '1',
+    #     'frequency': '/'.join([f'{n+1:01.0f}' for n in range(30)]),
+    #     'param': '251.140',
+    #     'stream': 'wave',
+    #     'time': '00:00:00/03:00:00/06:00:00/09:00:00/12:00:00/15:00:00/18:00:00/21:00:00',
+    #     'area': f'{lat[1]}/{lon[0]}/{lat[0]}/{lon[1]}', # north, west, south, east
+    #     'grid': f'{dlon}/{dlat}',
+    #     'type': 'an',
+    #     'format': 'netcdf',
+    #     }
 
     c.retrieve('reanalysis-era5-complete', cds_command, filename)
     return filename
 class ERA5(BoundaryReader):
+
+    def __init__(self):
+        self.dlon = 0.5
+        self.dlat = 0.5
+
     def convention(self) -> str:
         return 'Ocean'
 
     def get_coordinates(self, start_time) -> Tuple:
         """Reads first time instance of first file to get longitudes and latitudes for the PointPicker"""
-        point_list = self.get_restricted_area()._point_list()
+        restricted_area = self.get_restricted_area()
+        restricted_area.set_spacing(dlon=self.dlon, dlat=self.dlat)
+        point_list = restricted_area._point_list()
         lon_all = point_list[:,0]
         lat_all = point_list[:,1]
 
@@ -102,21 +143,31 @@ class ERA5(BoundaryReader):
             os.mkdir(temp_folder)
             print("Creating folder %s..." % temp_folder)
 
-        msg.plain("Removing old files from temporary folder...")
-        for f in glob.glob(f"{temp_folder}/EC_ERA5.nc"):
-            os.remove(f)
+        local_read = False
+
+        if not local_read:
+            msg.plain("Removing old files from temporary folder...")
+            for f in glob.glob(f"{temp_folder}/EC_ERA5.nc"):
+                os.remove(f)
 
         restricted_area = self.get_restricted_area()
+        lon = np.floor(np.array(restricted_area.lon_edges())/self.dlon)*self.dlon
+        lat = np.floor(np.array(restricted_area.lat_edges())/self.dlat)*self.dlat
 
-        nc_file = download_era5_from_cds(start_time, end_time,
-                                        lon=restricted_area.lon_edges(),
-                                        lat=restricted_area.lat_edges(),
-                                        dlon=restricted_area.dlon(),
-                                        dlat=restricted_area.dlat(),
-                                        folder=temp_folder)
+
+
+        if local_read:
+            nc_file = f'{temp_folder}/EC_ERA5.nc'
+        else:
+            nc_file = download_era5_from_cds(start_time, end_time,
+                                            lon=(lon[0]-self.dlon, lon[1]+self.dlon),
+                                            lat=(lat[0]-self.dlat, lat[1]+self.dlat),
+                                            dlon=self.dlon,
+                                            dlat=self.dlat,
+                                            folder=temp_folder)
 
         bnd_spec = xr.open_dataset(nc_file)
-
+        bnd_spec = bnd_spec.sortby("time")
         bnd_spec = renormalize_era5_spec(bnd_spec)
 
         lon, lat = np.meshgrid(bnd_spec.longitude.values, bnd_spec.latitude.values[::-1])
